Clean up, update on function and kernel tests
