+++ dotnet/src/SemanticKernel.UnitTests/Functions/KernelFunctionFromMethodTests1.cs
@@ -14,7 +14,7 @@ using Xunit;
 
 namespace SemanticKernel.UnitTests.Functions;
 
-public sealed class KernelFunctionTests2
+public sealed class KernelFunctionFromMethodTests1
 {
     private static readonly KernelFunction s_nopFunction = KernelFunctionFactory.CreateFromMethod(() => { });
 
@@ -24,7 +24,7 @@ public sealed class KernelFunctionTests2
     private static string s_expected = string.Empty;
     private static string s_actual = string.Empty;
 
-    public KernelFunctionTests2()
+    public KernelFunctionFromMethodTests1()
     {
         this._kernel = new Kernel();
         this._logger = new Mock<ILoggerFactory>();
